Moved total calc into utuility class and out of dashboard
